refactor: split tsdi into main and legacy build
The main tsdi build is not 12kb while the legacy one is around 26kb.
