+++ lib/index.ts
@@ -1,7 +1,1 @@
-export { Component as component, Component } from './component';
-export { Destroy as destroy, Destroy } from './destroy';
-export { External as external, External } from './external';
-export { Factory as factory, Factory } from './factory';
-export { Initialize as initialize, Initialize } from './initialize';
-export { Inject as inject, Inject } from './inject';
 export * from './tsdi';
