+++ package.json
@@ -59,7 +59,6 @@
     "husky": "1.3.1",
     "jest": "24.7.0",
     "lint-staged": "8.1.5",
-    "reflect-metadata": "0.1.13",
     "rimraf": "2.6.3",
     "rollup": "1.8.0",
     "rollup-plugin-analyzer": "3.0.0",
@@ -78,6 +77,7 @@
     "figlet": "^1.2.1",
     "lodash.kebabcase": "^4.1.1",
     "prettier": "1.16.4",
+    "reflect-metadata": "0.1.13",
     "resolve-from": "^4.0.0",
     "ts-morph": "^1.2.0",
     "typescript": "3.4.1",
