+++ tests/cyclic2.ts
@@ -1,4 +1,4 @@
-import { Component, Inject } from '../lib';
+import { Component, Inject } from '../lib/legacy';
 import { Cyclic1 } from './cyclic1';
 
 @Component()
