+++ tests/eager1.ts
@@ -1,4 +1,4 @@
-import { component, inject } from '../lib';
+import { component, inject } from '../lib/legacy';
 import { EagerComponent2 } from './eager2';
 
 @component({ eager: true })
