+++ tests/eager2.ts
@@ -1,4 +1,4 @@
-import { component } from '../lib';
+import { component } from '../lib/legacy';
 
 @component({ eager: true })
 export class EagerComponent2 {}
